move connect_[channel] to ConnectionFileMixin
