+++ IPython/kernel/manager.py
@@ -42,14 +42,6 @@ from .managerabc import (
 # Main kernel manager class
 #-----------------------------------------------------------------------------
 
-_socket_types = {
-    'hb' : zmq.REQ,
-    'shell' : zmq.DEALER,
-    'iopub' : zmq.SUB,
-    'stdin' : zmq.DEALER,
-    'control': zmq.DEALER,
-}
-
 class KernelManager(LoggingConfigurable, ConnectionFileMixin):
     """Manages a single kernel in a subprocess on this host.
 
@@ -143,54 +135,6 @@ class KernelManager(LoggingConfigurable, ConnectionFileMixin):
         # add kwargs last, for manual overrides
         kw.update(kwargs)
         return self.client_factory(**kw)
-
-    #--------------------------------------------------------------------------
-    # Connection info
-    #--------------------------------------------------------------------------
-
-    def _make_url(self, channel):
-        """Make a ZeroMQ URL for a given channel."""
-        transport = self.transport
-        ip = self.ip
-        port = getattr(self, '%s_port' % channel)
-
-        if transport == 'tcp':
-            return "tcp://%s:%i" % (ip, port)
-        else:
-            return "%s://%s-%s" % (transport, ip, port)
-
-    def _create_connected_socket(self, channel, identity=None):
-        """Create a zmq Socket and connect it to the kernel."""
-        url = self._make_url(channel)
-        socket_type = _socket_types[channel]
-        self.log.info("Connecting to: %s" % url)
-        sock = self.context.socket(socket_type)
-        if identity:
-            sock.identity = identity
-        sock.connect(url)
-        return sock
-
-    def connect_iopub(self, identity=None):
-        """return zmq Socket connected to the IOPub channel"""
-        sock = self._create_connected_socket('iopub', identity=identity)
-        sock.setsockopt(zmq.SUBSCRIBE, b'')
-        return sock
-
-    def connect_shell(self, identity=None):
-        """return zmq Socket connected to the Shell channel"""
-        return self._create_connected_socket('shell', identity=identity)
-
-    def connect_stdin(self, identity=None):
-        """return zmq Socket connected to the StdIn channel"""
-        return self._create_connected_socket('stdin', identity=identity)
-
-    def connect_hb(self, identity=None):
-        """return zmq Socket connected to the Heartbeat channel"""
-        return self._create_connected_socket('hb', identity=identity)
-
-    def connect_control(self, identity=None):
-        """return zmq Socket connected to the Heartbeat channel"""
-        return self._create_connected_socket('control', identity=identity)
 
     #--------------------------------------------------------------------------
     # Kernel management
